requirements.txt
fredapi>=0.5.0
matplotlib>=3.5.0
openpyxl>=3.0.0
pandas>=1.3.0
python-dotenv>=1.0.0
scipy>=1.7.0
seaborn>=0.12.0
yfinance>=0.2.0
